Update network settings and streamline docker services
The commit simplifies the docker-compose settings by renaming the network to net. Also, it optimizes and streamlines the docker services by removing unneeded deployments such as Zipkin and Jaeger for tempo and limits the receivers only to OTLP. Furthermore, datasource configurations for Grafana are added for Tempo and slight adjustments are made to the Nginx configuration.